refactor(ui): move unsaved_changes modal to dbflux_components::modals (#102)
Move ModalUnsavedChanges (with DirtySummaryEntry, UnsavedChangesOutcome,
UnsavedChangesRequest) from dbflux_ui into dbflux_components::modals/.
Replace the DocumentId import with dbflux_core::document_id::DocumentId
(no dbflux_ui edge). Tests relocated alongside. Shim at old path.
