+++ crates/dbflux_components/src/modals/mod.rs
@@ -4,6 +4,7 @@ pub mod drop_table;
 pub mod schema_drift;
 pub mod shell;
 pub mod tunnel_auth;
+pub mod unsaved_changes;
 
 pub use active_query::{ActiveQueryOutcome, ActiveQueryRequest, ActiveQueryTrigger, ModalActiveQuery};
 pub use delete_connection::{DeleteConnectionOutcome, DeleteConnectionRequest, ModalDeleteConnection};
@@ -11,3 +12,4 @@ pub use drop_table::{DropTableOutcome, DropTableRequest, ModalDropTable};
 pub use schema_drift::{ModalSchemaDrift, SchemaDriftContinue, SchemaDriftDismissed, SchemaDriftRefresh};
 pub use shell::{ModalShell, ModalVariant};
 pub use tunnel_auth::{ModalTunnelAuth, TunnelAuthOutcome, TunnelAuthRequest};
+pub use unsaved_changes::{DirtySummaryEntry, ModalUnsavedChanges, UnsavedChangesOutcome, UnsavedChangesRequest};
